Fixed tag margin aligments. APICheckboxList created as an independent component.
