+++ src/components/index.ts
@@ -25,3 +25,5 @@ export * from './subNav/SubNavEntry';
 export * from './subNav/SubNav';
 export * from './formField';
 export * from './appVersion/AppVersion';
+export * from './apiTags/ApiTags';
+export * from './apiCheckboxList/APICheckboxList';
